+++ tests/conftest.py
@@ -6,34 +6,16 @@ import typing
 
 import pytest
 
-
-def _generate_package(
-    config: dict[str, str], path: pathlib.Path
-) -> subprocess.CompletedProcess[str]:
-    """
-    Generate a project from the cookiecutter template.
-
-    Arguments:
-    ---------
-    config: dict
-        A dictionary with values for the cookiecutter template,
-        as defined in the cookiecutter.json
-    path: Path
-        Directory to create package in.
-
-    """
-    args = [f"{key}={val}" for key, val in config.items()]
-    cmd = ["cookiecutter", ".", "--no-input", "--output-dir", f"{path}"]
-    return subprocess.run(  # noqa: S603
-        cmd + args,
-        check=False,
-        shell=False,
-        capture_output=True,
-        text=True,
-    )
+from .helpers import DEFAULT_CONFIG, _generate_package  # type: ignore[import-not-found]
 
 
 @pytest.fixture
-def generate_package() -> typing.Callable:
+def generate_package(tmp_path: pathlib.Path) -> typing.Callable:
     """Generate project from cookiecutter template."""
-    return _generate_package
+
+    def wrapped_with_tmp_path(
+        config: dict[str, str] = DEFAULT_CONFIG,
+    ) -> tuple[subprocess.CompletedProcess, pathlib.Path]:
+        return _generate_package(config, tmp_path)
+
+    return wrapped_with_tmp_path
